requirements.txt
apturl==0.5.2
asciinema==2.0.2
asciitree==0.3.3
asgiref==3.7.2
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blessings==1.6
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.2.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.2.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
dnspython==1.16.0
docker==6.1.3
entrypoints==0.3
filelock==3.8.0
gpg===1.13.1-unknown
gpustat==0.6.0
httplib2==0.14.0
idna==2.8
Jinja2==2.10.1
kazam==1.5.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libiio==0.19
libvirt-python==6.1.0
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
mock==3.0.5
netifaces==0.10.4
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
ovs==2.13.8
packaging==20.3
pbr==5.4.5
pexpect==4.6.0
Pillow==7.0.0
pipenv==2022.8.31
platformdirs==2.5.2
protobuf==3.6.1
proton-client==0.7.1
protonvpn-cli==3.13.0
protonvpn-nm-lib==3.14.0
psutil==5.5.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
PyQt5==5.14.1
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
python-gnupg==0.4.5
python-xlib==0.23
pythondialog==3.4.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
QtPy==1.9.0
reportlab==3.5.34
requests==2.31.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sortedcontainers==2.1.0
speedtest-cli==2.1.3
sqlparse==0.4.4
ssh-import-id==5.10
systemd-python==234
typing-extensions==4.6.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==2.0.5
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.16.4
virtualenv-clone==0.5.7
wadllib==1.3.3
websocket-client==1.6.3
xkit==0.0.0
